@@ -31,51 +31,56 @@
         </developer>
     </developers>
 
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-parent</artifactId>
+        <version>2.5.3</version>
+    </parent>
+
     <properties>
-        <jdk.version>1.8</jdk.version>
+        <main.class>org.reactome.server.tools.event.exporter.Main</main.class>
+        <jdk.version>11</jdk.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <graph.core.version>1.1.27</graph.core.version>
-        <analysis.core.version>3.3.8</analysis.core.version>
-        <diagram.exporter.version>1.3.19-SNAPSHOT</diagram.exporter.version>
-        <reaction.exporter.version>1.1.6</reaction.exporter.version>
-        <fireworks.exporter.version>1.1.9</fireworks.exporter.version>
 
-        <junit.version>4.13.1</junit.version>
-        <itext.version>7.1.5</itext.version>
-        <spring.test.version>4.3.11.RELEASE</spring.test.version>
+        <graph.core.version>2.0.0</graph.core.version>
+        <analysis.core.version>3.4.0</analysis.core.version>
+        <diagram.exporter.version>2.0.0</diagram.exporter.version>
+        <reaction.exporter.version>1.2.0</reaction.exporter.version>
+        <fireworks.exporter.version>1.2.0</fireworks.exporter.version>
 
-        <maven.compiler.version>3.5.1</maven.compiler.version>
-        <maven.jar.version>2.6</maven.jar.version>
+        <junit.version>5.8.0-M1</junit.version>
+        <itext.version>7.1.15</itext.version>
+        <spring.test.version>5.3.6</spring.test.version>
 
+        <maven.compiler.version>3.8.1</maven.compiler.version>
+        <maven.jar.version>3.2.0</maven.jar.version>
+        <maven.resources.version>3.2.0</maven.resources.version>
+        <maven.assembly.version>3.3.0</maven.assembly.version>
+        <maven.surefire.version>3.0.0-M5</maven.surefire.version>
+        <maven.source.version>3.2.1</maven.source.version>
     </properties>
 
-
     <dependencies>
-
         <dependency>
             <groupId>org.reactome.server.graph</groupId>
             <artifactId>graph-core</artifactId>
             <version>${graph.core.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>analysis-core</artifactId>
             <version>${analysis.core.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>reaction-exporter</artifactId>
             <version>${reaction.exporter.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>diagram-exporter</artifactId>
             <version>${diagram.exporter.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.reactome.server.tools</groupId>
             <artifactId>fireworks-exporter</artifactId>
